chore: refactor event page to separate components
